step3 - web 컨트롤러, 뷰 생성 - 기본적인 동작 확인 - getTotalDistance 만듦 > 없이 출력하는 방법 필요
