+++ src/main/resources/templates/game.html
@@ -59,17 +59,12 @@ body {
 
         <div class="form-show-div" class="form-group">
           <table>
+              {{#cars}}
             <tr>
-              <td>박재성</td>
+              <td>{{name}}</td>
             </tr>
-            <tr>
-              <td>윤지수</td>
-            </tr>
-            <tr>
-              <td>정호영</td>
-            </tr>
+              {{/cars}}
           </table>
-
         </div>
         <form class="form-lotto" method="get" action="/result">
           Count 수 : <input type="number" name="turn">
