+++ src/main/resources/templates/result.html
@@ -1,76 +1,77 @@
 <!DOCTYPE html>
 <html>
 <head>
-<meta charset="UTF-8">
-<title>레이스!</title>
+    <meta charset="UTF-8">
+    <title>레이스!</title>
 
-<!-- Latest compiled and minified CSS -->
-<link rel="stylesheet"
-	href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
-	integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u"
-	crossorigin="anonymous">
+    <!-- Latest compiled and minified CSS -->
+    <link rel="stylesheet"
+          href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
+          integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u"
+          crossorigin="anonymous">
 
-<!-- Optional theme -->
-<link rel="stylesheet"
-	href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css"
-	integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp"
-	crossorigin="anonymous">
+    <!-- Optional theme -->
+    <link rel="stylesheet"
+          href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css"
+          integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp"
+          crossorigin="anonymous">
 
-<!-- Latest compiled and minified JavaScript -->
-<script
-	src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
-	integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
-	crossorigin="anonymous"></script>
-<style>
-body {
-	padding-top: 40px;
-	padding-bottom: 40px;
-	background-color: #eee;
-}
+    <!-- Latest compiled and minified JavaScript -->
+    <script
+            src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
+            integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
+            crossorigin="anonymous"></script>
+    <style>
+        body {
+            padding-top: 40px;
+            padding-bottom: 40px;
+            background-color: #eee;
+        }
 
-.form-lotto {
-	max-width: 1000px;
-	padding: 15px;
-	margin: 0 auto;
-}
+        .form-lotto {
+            max-width: 1000px;
+            padding: 15px;
+            margin: 0 auto;
+        }
 
-.form-lotto .form-control {
-	position: relative;
-	height: auto;
-	-webkit-box-sizing: border-box;
-	-moz-box-sizing: border-box;
-	box-sizing: border-box;
-	padding: 10px;
-	font-size: 16px;
-}
+        .form-lotto .form-control {
+            position: relative;
+            height: auto;
+            -webkit-box-sizing: border-box;
+            -moz-box-sizing: border-box;
+            box-sizing: border-box;
+            padding: 10px;
+            font-size: 16px;
+        }
 
-.submit-button {
-	margin-top: 10px;
-}
-</style>
+        .submit-button {
+            margin-top: 10px;
+        }
+    </style>
 
 </head>
 <body>
 
-  <div class="container">
+<div class="container">
     <div class="row">
 
-      <div class="col-md-12">
+        <div class="col-md-12">
 
-        <h2 class="text-center">박진감 넘치는 자동차 레이싱 게임 with Pobi</h2>
-        <div class="form-show-div" class="form-group">
-          <h3>결과입니다</h3>
-          (이름마다 띄어쓰기로 구분해 주세요.)
-          <div id="standings">박재성 : &nbsp;&nbsp;&nbsp;&#128652;</div>
-          <div id="standings">윤지수 : &nbsp;&nbsp;&#128652;</div>
-          <div id="standings">정호영 : &nbsp;&nbsp;&nbsp;&#128652;</div>
-          <div id="standings">김정 : &nbsp;&#128652;</div>
-          <br />
-          <div>박재성, 정호영이 우승했습니다.</div>
+            <h2 class="text-center">박진감 넘치는 자동차 레이싱 게임 with Pobi</h2>
+            <div class="form-show-div" class="form-group">
+                <h3>결과입니다</h3>
+                (이름마다 띄어쓰기로 구분해 주세요.)
+                {{#cars}}
+                <div id={{name}}>{{name}} : {{totalDistance}}&#128652;</div>
+                {{/cars}}
+                <br />
+
+                <div>{{winners}} 이 우승했습니다.</div>
+
+            </div>
         </div>
-      </div>
     </div>
-  </div>
+</div>
 
 </body>
 <script src="/js/race.js">
